Vacaciones Fraccionadas Backend 100%

+++ database/migrations/2018_09_19_214600_create_vacaciones_table.php
@@ -17,7 +17,7 @@ class CreateVacacionesTable extends Migration
             $table->increments('id');
             $table->integer('trabajador_id')->unsigned();
             $table->integer('a_servicio');
-            $table->integer('dias_disfrute');
+            $table->float('dias_disfrute');
             $table->integer('dias_feriados');
             $table->integer('dias_descanso');
             $table->string('tipo');
